Get Methods
Ive added get methods for the server.js. The first removes the element in a array called allNotes that will be created to take in notes inpuitted by application. This is because the first value of any array is 0 and since this doesnt represent any data i have remove it so it starts with 1 which will be the note inputted. The rest join paths of the index.html and notes.html and set a wildcardcalled cath-all to serve as a fallback page.