Delete Note Function, app.delete, and app.listen
THis is my current code for deleting a note after its been created and having the express app to listen on my specific port with a console.log. The deleteNote function iterates over each note in the allNotesArray and finds a note with an ID that matches the id parameter which removes the note from the array. This then writes the updated array to the json file and breaks out of the loop. The app.delete sets up a route where :id is a parameter that represents the ID of the note to be deleted. When accessed the deleteNote function is called and sends a json object with a value of true back in response.